src/auth/pyproject.toml
[project]
name = "auth"
version = "1.5.0"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.4",
    "argon2-cffi>=25.1.0",
    "asyncpg==0.30.0",
    "blake3>=1.0.8",
    "cashews[redis]==7.4.1",
    "greenlet>=3.2.4",
    "grpcio==1.74.0",
    "grpcio-tools==1.74.0",
    "httpagentparser==1.9.5",
    "jwskate==0.12.2",
    "py-async-grpc-prometheus==0.2.0",
    "sqlalchemy==2.0.43",
    "uvicorn[standard]==0.35.0",
]

[dependency-groups]
test = [
    "coverage==7.10.3",
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
    "pytest-cov==6.2.1",
]
